fix: hold master's merged-in code to this branch's ruff 0.16 + pyright standards
The merge in the previous commit was clean but left CI red, because master's 86 commits were
written under the OLD ruff config (`ruff>=0.3`, so effectively ruff's pre-0.16 defaults) and had
never been checked against the extended rule set this branch adopts.

Measured at CI's exact scope (`ruff check osmose/ ui/ tests/ scripts/ app.py`) with the PINNED
ruff 0.16.0, not a locally-installed one:

    705ce53 (pre-merge)  -> All checks passed
    0be59d9 (merge)      -> 92 errors

86 were auto-fixable. The remaining 6 needed judgement, and two of them are NOT defects:

  - `osmose/feedback.py` DTZ005 (`datetime.now()` without tz) is SUPPRESSED, not changed. A
    tz-aware timestamp would alter the `ts` field of the stored feedback record, which is live in
    production since 2026-09-15 and already written to disk. That is a data-format decision, not a
    lint fix.
  - `scripts/c3_growth_deficit_diagnosis.py` PLR0124 (`x == x`) is the NaN idiom -- NaN is not
    equal to itself, so the expression prints "--" for NaN. Suppressed with that reason.

The other four are mechanical and behaviour-preserving: a redundant `int(round(...))`, a dict
iterated with `.items()` when only values are used, a percent-format string, and a mutable class
attribute needing `ClassVar`.

Also fixes the 8 pyright errors the merge introduced, all in `ui/components/feedback_modal.py`
(feedback v2 code from master, never type-checked under this branch's config). Measured the same
way, with `PYRIGHT_PYTHON_FORCE_VERSION=latest` as this repo's own pyproject comment instructs:

    705ce53 -> 21 errors      0be59d9 -> 29 errors      now -> 21 errors

Fixed at the type level rather than suppressed:
  - `tag.children` is a union and only `Tag` carries `.attrs`; the `getattr(child, "name")` guard
    was an unnarrowed attribute access that happened to work. `isinstance(child, Tag)` narrows it
    and is the stronger check.
  - the notification lambdas return shiny's value (a str id), so the list is
    `Callable[[], object]`, not `Callable[[], None]`.
  - `_notify`'s `type` parameter is shiny's `Literal["default","message","warning","error"]`.
  - `_classify_and_consume` returns `str | None`; REJECT always carries a message, but fall back
    rather than pass None into a `str` parameter.

NOT addressed, and NOT caused by the merge: the Docs and Visual workflows were already failing on
705ce53 (`osmose/calibration/stability.py`'s docstring trips two sphinx warnings under `-W`, and
that file is untouched by the merge).

+++ scripts/c3_juvenile_boost_literature_check.py
@@ -63,7 +63,7 @@ from scipy.stats import pearsonr, spearmanr
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
 
-from osmose.calibration.bioen_offline import (  # noqa: E402
+from osmose.calibration.bioen_offline import (
     BioenFixed,
     c_m_from_share,
     fit_species,
@@ -71,9 +71,9 @@ from osmose.calibration.bioen_offline import (  # noqa: E402
     solve_tp,
     vbgf_weight,
 )
-from osmose.config import OsmoseConfigReader  # noqa: E402
-from osmose.demo import osmose_demo  # noqa: E402
-from osmose.engine.config import EngineConfig  # noqa: E402
+from osmose.config import OsmoseConfigReader
+from osmose.demo import osmose_demo
+from osmose.engine.config import EngineConfig
 
 NDT = 24
 BETA_EMPIRICAL = 0.75  # Kiorboe & Hirst 2014: ingestion ~ w^(3/4)
@@ -125,7 +125,7 @@ def main() -> int:
     for tg in targets:
         res = fit_species(tg, fx, ndt=NDT, juvenile_boost=True)
         j = res.juvenile_boost
-        n_steps = int(round(tg.lifespan_years * NDT))
+        n_steps = round(tg.lifespan_years * NDT)
         t_p = solve_tp(tg.t_opt, fx)
 
         # The fitted model's own weight path, including the boost it chose.
